agent/requirements.txt
google-adk[a2a]
python-dotenv
httpx
